the_balance_of_a_bank_acount_after_two_deposits_is_the_sum_of_the_deposits

+++ src/main/java/bank/Amount.java
@@ -3,9 +3,9 @@ package bank;
 import java.util.Objects;
 
 public class Amount {
-    private final int value;
+    private final long value;
 
-    public Amount(int value) {
+    public Amount(long value) {
         this.value = value;
     }
